requirements.txt
langchain
python-dotenv
langchain-community
bs4 #beautifulSoup
Wikipedia
langchain-text-splitters
langchain-openai
chromadb
sentence-transformers
langchain_huggingface
faiss-cpu
langchain_chroma
beautifulsoup4
streamlit